weightage u8 (#2992)

+++ lib/ain-ocean/src/model/oracle_price_aggregated.rs
@@ -19,7 +19,7 @@ pub struct OraclePriceAggregated {
 #[serde(rename_all = "camelCase")]
 pub struct OraclePriceAggregatedAggregated {
     pub amount: String,
-    pub weightage: i32,
+    pub weightage: u8,
     pub oracles: OraclePriceAggregatedAggregatedOracles,
 }
